+++ apps/atlas/src/components/minigame/realtime/Chat.tsx
@@ -6,9 +6,6 @@ import { useAccount } from '@starknet-react/core';
 import type { Types } from 'ably';
 import classNames from 'classnames';
 import React, { useEffect, useRef, useState } from 'react';
-import useGameStatus from '@/hooks/desiege/useGameStatus';
-import useGameVariables from '@/hooks/desiege/useGameVariables';
-import { useTokenBalances } from '@/hooks/desiege/useTokenBalances';
 
 import TokenLabel, {
   DarkGradient,
@@ -37,31 +34,7 @@ const ChatComponent = (props: ChatComponentProps) => {
 
   const { address } = useAccount();
 
-  const gameVars = useGameVariables();
-  const getGameStatus = useGameStatus({
-    gameIdx: gameVars.data?.gameIdx,
-  });
-
-  // If the game is active, use the current game index.
-  // Else the balls should show the total minted for the next game
-  let gameIdx;
-  if (gameVars.data && getGameStatus.data) {
-    if (getGameStatus.data == 'active') {
-      gameIdx = gameVars.data.gameIdx;
-    } else {
-      gameIdx = gameVars.data.gameIdx + 1;
-    }
-  }
-
-  const tokenBalance = useTokenBalances({
-    gameIdx,
-  });
-  const resolvedChannelName =
-    messagesFilterOnlySide && tokenBalance.side
-      ? `${props.channelName}-${tokenBalance.side}`
-      : props.channelName;
-  ``;
-  const [channel] = useChannel(resolvedChannelName, (message) => {
+  const [channel] = useChannel(props.channelName, (message) => {
     // 200 is the max number of messages to keep in the chat
     setMessages((msgs) => [...msgs.slice(-199), message]);
   });
@@ -86,11 +59,9 @@ const ChatComponent = (props: ChatComponentProps) => {
   }, [receivedMessages]);
 
   const sendChatMessage = (messageText) => {
-    console.log('Sending chat message for side', tokenBalance.side);
     channel.publish({
       name: 'chat-message',
       data: {
-        side: tokenBalance.side,
         body: messageText,
         address,
       },
@@ -141,28 +112,6 @@ const ChatComponent = (props: ChatComponentProps) => {
             <div className="self-center w-2 h-2 mr-2 bg-green-500 rounded-full"></div>
             {presenceData.length} online
           </div>
-
-          {tokenBalance.side && (
-            <button
-              onClick={() => setMessagesFilterOnlySide((prev) => !prev)}
-              className={classNames(
-                'bg-gradient-to-t ml-2 rounded-sm px-1 text-sm',
-                messagesFilterOnlySide && tokenBalance.side == 'light'
-                  ? LightGradient
-                  : null,
-                messagesFilterOnlySide && tokenBalance.side == 'dark'
-                  ? DarkGradient
-                  : null
-              )}
-            >
-              {messagesFilterOnlySide ? (
-                <SolidCircleIcon className="inline-block w-3" />
-              ) : (
-                <CheckCircleIcon className="inline-block w-3" />
-              )}{' '}
-              {tokenBalance.side?.toUpperCase()} Only
-            </button>
-          )}
         </span>
       </div>
       <div className="h-32 p-2 overflow-y-scroll bg-black border rounded-md card max-h-32">
