+++ apps/atlas/src/components/tables/Transactions.tsx
@@ -1,11 +1,11 @@
 import { Button } from '@bibliotheca-dao/ui-lib/base';
 import DragIcon from '@bibliotheca-dao/ui-lib/icons/drag.svg';
-import { useTransactionManager } from '@starknet-react/core';
+import type { ManagedTransaction } from '@starknet-react/core';
+import { useTransactionManager, useTransactions } from '@starknet-react/core';
 import type { Identifier, XYCoord } from 'dnd-core';
 import Link from 'next/link';
 import type { ReactElement } from 'react';
 import { useEffect, useRef } from 'react';
-import { useDrag, useDrop } from 'react-dnd';
 import type { Status, TransactionStatus } from 'starknet';
 import { twMerge } from 'tailwind-merge';
 import type { ENQUEUED_STATUS } from '@/constants/index';
@@ -19,16 +19,11 @@ interface Metadata {
   action: string;
   multicalls?: any;
 }
-interface EnqueuedOrReceivedTransaction {
-  metadata?: Metadata;
-  status: Status | TransactionStatus | typeof ENQUEUED_STATUS;
-  transactionHash?: string;
-}
+
 interface TxCartItem {
   index: number;
-  transaction: EnqueuedOrReceivedTransaction;
-  onRemove?: () => void;
-  onReorder?: (dragIndex: number, hoverIndex: number) => void;
+  transaction: ManagedTransaction<Metadata>;
+  status?: Status | TransactionStatus | typeof ENQUEUED_STATUS;
 }
 
 const STYLES = {
@@ -85,102 +80,21 @@ export const TxCartItem = (props: TxCartItem) => {
     );
   }
 
-  const [{ handlerId }, drop] = useDrop<
-    DragTx,
-    void,
-    { handlerId: Identifier | null }
-  >({
-    accept: dndTypes.TX,
-    collect(monitor) {
-      return {
-        handlerId: monitor.getHandlerId(),
-      };
-    },
-    hover(item: DragTx, monitor) {
-      if (!ref.current) {
-        return;
-      }
-      const dragIndex = item.index;
-      const hoverIndex = props.index;
-
-      // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
-        return;
-      }
-
-      // Determine rectangle on screen
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-
-      // Get vertical middle
-      const hoverMiddleY =
-        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset();
-
-      // Get pixels to the top
-      const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
-
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-
-      // Dragging downwards
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return;
-      }
-
-      // Dragging upwards
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return;
-      }
-
-      // Time to actually perform the action
-      props.onReorder && props.onReorder(dragIndex, hoverIndex);
-
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      item.index = hoverIndex;
-    },
-  });
-  const [{ isDragging }, drag] = useDrag({
-    type: dndTypes.TX,
-    item: () => {
-      return {
-        transactionHash: props.transaction.transactionHash,
-        index: props.index,
-      };
-    },
-    collect: (monitor: any) => ({
-      isDragging: monitor.isDragging(),
-    }),
-  });
-
-  // no dnd if tx is in status tab
-  props.onRemove && drag(drop(ref));
-
   return (
     <div
       ref={ref}
-      data-handler-id={handlerId}
       className={twMerge(
-        STYLES.status[props.transaction.status],
-        `rounded-xl flex p-4 w-full mb-2  border card`,
-        isDragging ? 'opacity-0' : 'opacity-100'
+        STYLES.status[props.status || 0],
+        `rounded-xl flex p-4 w-full mb-2  border card`
       )}
     >
-      {props.onRemove && (
-        <DragIcon className="fill-current w-5 -translate-x-1.5 cursor-grab" />
-      )}
       <div className="flex flex-wrap w-full p-1 rounded bg-black/19">
         <div className="flex justify-between w-full pb-2 mb-2 border-b border-white/20">
           <h5 className="self-center uppercase">
-            {FORMATED_STATUS.status[props.transaction.status]}
+            {FORMATED_STATUS.status[props.status || 0]}
           </h5>
           <div className="self-center ">
-            {props.transaction.transactionHash ? (
+            {props.transaction.hash ? (
               <Button
                 target={'_blank'}
                 size="xs"
@@ -188,23 +102,12 @@ export const TxCartItem = (props: TxCartItem) => {
                 rel="noreferrer noopener"
                 href={
                   // TODO: use network aware link using @/util/blockExplorer
-                  'https://goerli.voyager.online/tx/' +
-                  props.transaction.transactionHash
+                  'https://testnet.starkscan.co/tx/' + props.transaction.hash
                 }
               >
-                See on Voyager <ExternalLink className="inline-block w-4" />
+                See on StarkScan <ExternalLink className="inline-block w-4" />
               </Button>
             ) : null}
-            {props.onRemove && (
-              <Button
-                size="xs"
-                variant="outline"
-                texture={false}
-                onClick={props.onRemove}
-              >
-                Remove
-              </Button>
-            )}
           </div>
         </div>
 
@@ -215,10 +118,7 @@ export const TxCartItem = (props: TxCartItem) => {
               const renderConfig = getTxRenderConfig(tx);
               if (renderConfig) {
                 return (
-                  <div
-                    className="py-2"
-                    key={`${props.transaction.transactionHash}:${i}`}
-                  >
+                  <div className="py-2" key={`${props.transaction.hash}:${i}`}>
                     <h5 className="uppercase">{renderConfig.title}</h5>
                     <p className="text-lg font-semibold opacity-75">
                       {renderConfig.description}
@@ -239,7 +139,9 @@ export const TxCartItem = (props: TxCartItem) => {
 };
 
 export const TransactionCartTable = () => {
-  const { transactions } = useTransactionManager();
+  const { hashes, transactions } = useTransactionManager<Metadata>();
+  const transactionStatus = useTransactions({ hashes });
+
   return (
     <div className="flex flex-col-reverse flex-wrap w-full">
       {transactions.map((a, index) => {
