+++ apps/atlas/src/shared/squad/Squad.tsx
@@ -5,7 +5,7 @@
 // import { ArmoryBuilder } from '@/components/tables/Armory';
 // import { Squad } from '@/constants/index';
 // import { findResourceById } from '@/constants/resources';
-// import { useTransactionQueue } from '@/context/TransactionQueueContext';
+// import { useCommandList} from '@/context/CommandListContext';
 // import type { GetRealmQuery } from '@/generated/graphql';
 // import { useGetTroopStatsQuery } from '@/generated/graphql';
 // import useCombat, { createCall } from '@/hooks/settling/useCombat';
