+++ apps/desiege-silo/hooks/desiege/useTotalMinted.tsx
